pyproject.toml
[tool.poetry]
name = "hot_100"
version = "0.1.0"
description = ""
authors = ["Sean <seanh@zillowgroup.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
